+++ CHANGELOG.md
@@ -7,9 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Allow setting the `traverse` state of an `entity.component_tags[type][entity](traverse=...)` attribute.
+
 ### Changed
 
 - Updated deprecations to use [PEP 702](https://peps.python.org/pep-0702/).
+
+### Fixed
+
+- Fixed `.clear` methods for `entity.components` and `entity.component_relations`.
 
 ## [5.3.0] - 2025-03-08